LayoutTests/platform/ios/imported/w3c/web-platform-tests/editing/other/input-in-text-control-which-is-also-editing-host.tentative_textcontrol=textarea-expected.txt



FAIL User typing in <textarea contenteditable> should update the value assert_equals: User typing in <textarea contenteditable> should update the value: The text control value should be updated expected "abc" but got ""
PASS execCommand("insertText") in <textarea contenteditable> should update the value
FAIL User typing in <textarea contenteditable> should update the value (became an editing host during focused) assert_equals: User typing in <textarea contenteditable> should update the value (became an editing host during focused): The text control value should be updated expected "abc" but got ""
PASS execCommand("insertText") in <textarea contenteditable> should update the value (became an editing host during focused)
